Ui events now take in a generic type on success or submit.
This is so in the controller, on a service success, it can pass in the results of the service,
and so the ui event can act according to the service result.

What we're going to aim at is the add thread result then allowing a ui event to open a new thread with the
new thread id.